move utils under helpers folder
